range moved back and shared cancelled
