+++ parser_library/include/workspace_manager.h
