+++ parser_library/src/context/hlasm_statement.h
@@ -17,7 +17,7 @@
 
 #include <memory>
 #include <vector>
-#include "../include/shared/range.h"
+#include "include/range.h"
 
 namespace hlasm_plugin {
 namespace parser_library {
